chore(changesets): bump packages version (#6)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.

# Releases
## @jsxcss/core@1.2.0

### Minor Changes

- ad3864c: feat(emotion): add responsive mediaQuery prop

## @jsxcss/emotion@1.2.0

### Minor Changes

- ad3864c: feat(emotion): add responsive mediaQuery prop

### Patch Changes

- Updated dependencies [ad3864c]
    - @jsxcss/core@1.2.0

## @jsxcss/devtool@1.2.0

---------

Co-authored-by: Jonghyeon Ko <manudeli.ko@gmail.com>